Update and rename ufo_sightings.R to outlining_drivers_for_UFO_sightings.R
Part 2 expands on Drew Conway's code (Part 1) by clustering the mass sighting events in categories to conduct further analysis. The focus will be on the 'Short description' or Shape parameter of the data with the analysis aiding to determine its relation to the actual driver of the UFO sighting event.